pyproject.toml
[tool.poetry]
name = "investment_portfolio"
version = "0.1.0"
description = ""
authors = ["Griffin Knipe <knipegp@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
pyportfolioopt = "^1.5.1"
yfinance = "^0.1.70"
requests-cache = "^0.9.3"
PyYAML = "^6.0"
marshmallow = "^3.14.1"
scikit-learn = "^1.0.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^22.1.0"
pylint = "^2.12.2"
mypy = "^0.931"

[tool.poetry.scripts]
do_it = 'investment_portfolio.main:main'

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
